Refactored Schedule module
