+++ src/guis/zeale/guis/schedule_module/ScheduleModule.fxml
@@ -6,14 +6,10 @@
 <?import javafx.scene.layout.*?>
 <?import javafx.scene.layout.AnchorPane?>
 
-<AnchorPane prefHeight="1080.0" prefWidth="1920.0" styleClass="background" stylesheets="@css/stylesheet.css" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1">
+<AnchorPane prefHeight="1080.0" prefWidth="1920.0" styleClass="background" stylesheets="@../css/stylesheet.css" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1">
    <children>
       <ListView fx:id="eventList" layoutX="14.0" layoutY="344.0" prefHeight="736.0" prefWidth="1920.0" styleClass="rgba-background" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="344.0" />
-      <Group fx:id="eventUnselectedGroup">
-         <children>
-            <Button focusTraversable="false" layoutX="73.0" layoutY="72.0" mnemonicParsing="false" onAction="#addEvent" styleClass="pop-button" text="Add Event" />
-         </children>
-      </Group>
+      <Group fx:id="eventUnselectedGroup" />
       <Group fx:id="eventSelectedGroup" layoutX="-126.0" visible="false">
          <children>
             <Button focusTraversable="false" layoutX="196.0" layoutY="142.0" mnemonicParsing="false" styleClass="pop-button" text="Edit Event" />
@@ -21,5 +17,6 @@
             <Button focusTraversable="false" layoutX="189.0" layoutY="108.0" mnemonicParsing="false" styleClass="pop-button" text="Rename Event" />
          </children>
       </Group>
+      <Button focusTraversable="false" layoutX="78.0" layoutY="146.0" mnemonicParsing="false" onAction="#addEvent" prefHeight="34.0" prefWidth="97.0" style="-fx-font-size: 18;" styleClass="pop-button" text="New" />
    </children>
 </AnchorPane>
